Changed the api

+++ app/api/pools/route.js
@@ -2,7 +2,29 @@ import { NextResponse } from "next/server"
 import prisma from "../../libs/prismadb";
 
 export async function GET(request) {
-  const pools = await prisma.pool.findMany({
+  const id = request.nextUrl.searchParams.get("id");
+
+  if (!id) {
+    const pools = await prisma.pool.findMany({
+      include: {
+        teams: true,
+        matches: {
+          include: {
+            team1: true,
+            team2: true,
+            workTeam: true,
+            sets: true
+          }
+        },
+      }
+    })
+  
+    return NextResponse.json(pools)
+  }
+  const pool = await prisma.pool.findUnique({
+    where: {
+      id: id
+    },
     include: {
       teams: true,
       matches: {
@@ -16,7 +38,10 @@ export async function GET(request) {
     }
   })
 
-  return NextResponse.json(pools)
+  if (!pool) {
+    return new NextResponse("Need Pool Name", { status: 400 })
+  }
+  return NextResponse.json(pool)
 }
 
 export async function POST(request, { params }) {
@@ -36,10 +61,21 @@ export async function POST(request, { params }) {
 }
 
 export async function DELETE(request) {
+  const id = request.nextUrl.searchParams.get("id");
 
-  const pools = await prisma.pool.deleteMany({})
+  if (!id) {
+    const pools = await prisma.pool.deleteMany({})
 
-  return NextResponse.json(pools)
+    return NextResponse.json(pools)
+  }
+
+  const pool = await prisma.pool.delete({
+    where: {
+      id: params.id
+    }
+  })
+
+  return NextResponse.json(pool)
 
 }
